updated comments to have polymorphic relations

+++ database/migrations/2021_01_05_081915_add_cascade_delete_to_comments_table.php
@@ -14,14 +14,14 @@ class AddCascadeDeleteToCommentsTable extends Migration
     public function up()
     {
         Schema::table('comments', function (Blueprint $table) {
-            if (env('DB_CONNECTION') !== 'sqlite_testing') {
+            if (env('DB_CONNECTION') !== 'mysql_testing') {
                 $table->dropForeign(['blog_post_id']);
                 $table->dropColumn('blog_post_id');
             }
         });
 
         Schema::table('comments', function (Blueprint $table) {
-            if (env('DB_CONNECTION') !== 'sqlite_testing') {
+            if (env('DB_CONNECTION') !== 'mysql_testing') {
                 $table->unsignedInteger('blog_post_id');
                 $table->foreign('blog_post_id')
                     ->references('id')
@@ -40,8 +40,10 @@ class AddCascadeDeleteToCommentsTable extends Migration
     {
 
         Schema::table('comments', function (Blueprint $table) {
-            $table->dropForeign(['blog_post_id']);
-            $table->dropIfExists('blog_post_id');
+            if (Schema::hasColumn('comments', 'blog_post_id')) {
+                $table->dropForeign(['blog_post_id']);
+                $table->dropIfExists('blog_post_id');
+            }
         });
 
         Schema::table('comments', function (Blueprint $table) {
